Highlight cell
DIstinguish the cell selected by the user. As it is not possible to change the opacity or color of a specific cell of a trace, it is required to draw a new plot on top of the existing one. It is required to delete the trace afterwards to clean the previous highlight.