Animation when loading data
When the "Load data" button is clicked, the heatmap tab is selected and a loading spinner appears on top until the plot is completely loaded.
It was ecessary to do some modifications on Tabs to bind the function and be able to trigger the selection from the App component. Also the loading status is bind between the App and Heatmap components.