Display performance stat with 3 differentiated colours
The stat related to the performance of the map (FAST in all our cases), is displayed in 3 different colours; green, yellow or red. The value assigned depends on the min and max possible valued found in the map and shows an approximation reference of the good, average or bad performance of that agent.