Fix video refresh issue
Make sure that when a new cell of the heatmap is selected, the gameplay video corresponds to the one of the agent selected. The VideoPlayer component used was not being updated when the src changed, so a new one of our own has been created, It is possible to pause and play the video.